# openapi/api-spec.openapi.yaml
openapi: 3.1.2
info:
  title: Transport Booking API
  version: 1.0.0
  description: >
    ## Overview

    API for booking vehicles owned by drivers, used by clients.

    ## Authentication

    Methods requiring authentication should send the header `Authentication: Bearer <token>`.
    The <token> is a jwt and may be obtained by making a request to any endpoint with the header
    `Authentication: Basic <credentials>`, where `<credentials>` is the Base64 encoding of
    `<username>:<password>`. If the credentials are correct then the server will reply with the
    normal endpoint's response and a jwt in the header `vango-auth-token: <token>` and
    `vango-refresh-token: <refresh-token>`.
servers:
  - url: http://localhost:8080/api
    description: Development Server
...
paths:
  /clients:
    post:
      summary: Create a new client
      description: Registers a new client account in the system.
      operationId: createClient
      tags:
        - Clients
      requestBody:
        required: true
        description: >
          Client information required to create a new client account.
          The email address must be unique.
        content:
          # TODO: use custom versioned mimetype?
          application/json:
            schema:
              $ref: "schemas/requests/client-create.yaml"
      responses:
        "201":
          description: Client created
          headers:
            Location:
              description: URI of the created client
              schema:
                type: string
                example: "api/clients/13"
          content:
            application/json:
              schema:
                $ref: "schemas/responses/client.yaml"
        "400":
          $ref: "schemas/errors/bad-request.yaml"
        "409":
          $ref: "schemas/errors/conflict.yaml"
  /clients/{clientId}:
    get:
      parameters:
        - $ref: "#/components/parameters/ClientIdPath"
      summary: Get information of a Client
      description: Given a client ID, obtain the information of a Client.
      operationId: getClientById
      tags:
        - Clients
      responses:
        "200":
          description: Information of a client
          content:
            application/json:
              schema:
                $ref: "schemas/responses/client.yaml"
        "401":
          $ref: "schemas/errors/unauthorized.yaml"
        "404":
          $ref: "schemas/errors/not-found.yaml"
    patch:
      parameters:
        - $ref: "#/components/parameters/ClientIdPath"
      summary: Update client information
      description: Given a client ID, update the specified information of a Client.
      operationId: updateClientById
      tags:
        - Clients
      requestBody:
        required: true
        description: Fields that may be updated for an existing client.
        content:
          application/json:
            schema:
              $ref: "schemas/requests/client-update.yaml"
      responses:
        "200":
          description: Client updated successfully
          content:
            application/json:
              schema:
                $ref: "schemas/responses/client.yaml"
        "400":
          $ref: "schemas/errors/bad-request.yaml"
        "401":
          $ref: "schemas/errors/unauthorized.yaml"
        "403":
          $ref: "schemas/errors/forbidden.yaml"
        "404":
          $ref: "schemas/errors/not-found.yaml"
        "409":
          $ref: "schemas/errors/conflict.yaml"
  /clients/{clientId}/profile-picture:
    get:
      parameters:
        - $ref: "#/components/parameters/ClientIdPath"
      summary: Get the profile picture.
      description: Get the profile picture of the client id.
      operationId: getClientProfilePicture
      tags:
        - Clients
      responses:
        "200":
          description: Client's profile picture
          $ref: "schemas/image.yaml"
        "401":
          $ref: "schemas/errors/unauthorized.yaml"
        "404":
          $ref: "schemas/errors/image-not-found.yaml"
    post:
      parameters:
        - $ref: "#/components/parameters/ClientIdPath"
      tags:
        - Clients
      summary: Upload client profile picture
      description: Uploads or replaces the profile picture for a client (format PNG or JPG).
      operationId: uploadClientProfilePicture
      requestBody:
        required: true
        description: Image file to use as the client’s profile picture
        $ref: "schemas/image.yaml"
      responses:
        "201":
          description: Profile picture uploaded
          headers:
            Location:
              description: URI of the uploaded image
              schema:
                type: string
                example: "api/client/42/profile-picture"
        "401":
          $ref: "schemas/errors/unauthorized.yaml"
        "403":
          $ref: "schemas/errors/forbidden.yaml"
        "415":
          $ref: "schemas/errors/unsupported-media-type.yaml"
...
components:
  parameters:
    ClientIdPath:
      name: clientId
      in: path
      required: true
      description: ID of the client
      schema:
        type: integer
...